+++ lib/glorbo/cli/scaffold/project.ex
@@ -85,7 +85,7 @@ defmodule Glorbo.CLI.Scaffold.Project do
   end
 
   defp glorbo_home do
-    System.get_env("GLORBO_HOME") || Glorbo.Filesystem.Hierarchy.default_root()
+    Glorbo.Filesystem.Hierarchy.home_root()
   end
 
   defp usage do